+++ clare/clare/retry/__init__.py
@@ -2,8 +2,8 @@
 
 from . import stop_strategies
 from . import wait_strategies
-from .retry_policy import RetryPolicy
+from .retry_policy_builder import RetryPolicyBuilder
 
-__all__ = ['RetryPolicy',
+__all__ = ['RetryPolicyBuilder',
            'stop_strategies',
            'wait_strategies']