+++ clare/clare/retry/tests/test_retry_policy.py
@@ -3,7 +3,7 @@
 import mock
 from nose.tools import assert_equal, assert_greater
 
-from .. import RetryPolicy, stop_strategies, wait_strategies
+from .. import RetryPolicyBuilder, stop_strategies, wait_strategies
 
 
 class MockException(Exception):
@@ -39,33 +39,37 @@ class TestRetryPolicy(object):
         self.handled_exceptions = (MockException,)
 
     def test_execute_stop_after_success(self):
-        retry_policy = RetryPolicy(
-            stop_strategy=stop_strategies.AfterNever(),
-            wait_strategy=wait_strategies.Fixed(wait_time=0))
+        retry_policy = RetryPolicyBuilder() \
+            .with_stop_strategy(stop_strategies.AfterNever()) \
+            .with_wait_strategy(wait_strategies.Fixed(wait_time=0)) \
+            .build()
         retry_policy.execute(callable=self.service.call)
         assert_equal(self.service.call_count, 1)
 
     def test_execute_wait(self):
         _sleep = mock.MagicMock()
-        retry_policy = RetryPolicy(
-            stop_strategy=stop_strategies.AfterAttempt(maximum_attempt=2),
-            wait_strategy=wait_strategies.Fixed(wait_time=None),
-            handled_exceptions=self.handled_exceptions)
+        retry_policy = RetryPolicyBuilder() \
+            .with_stop_strategy(stop_strategies.AfterAttempt(maximum_attempt=2)) \
+            .with_wait_strategy(wait_strategies.Fixed(wait_time=None)) \
+            .continue_on_exception(MockException) \
+            .build()
         retry_policy.execute(callable=self.service.call_and_raise,
                              _sleep=_sleep)
         _sleep.assert_called()
 
     def test_execute_successful_attempt_returns_result(self):
-        retry_policy = RetryPolicy(
-            stop_strategy=stop_strategies.AfterNever(),
-            wait_strategy=wait_strategies.Fixed(wait_time=0))
+        retry_policy = RetryPolicyBuilder() \
+            .with_stop_strategy(stop_strategies.AfterNever()) \
+            .with_wait_strategy(wait_strategies.Fixed(wait_time=0)) \
+            .build()
         result = retry_policy.execute(callable=self.service.call_and_return)
         assert_equal(result, 'foo')
 
     def test_execute_continue_on_handled_exceptions(self):
-        retry_policy = RetryPolicy(
-            stop_strategy=stop_strategies.AfterAttempt(maximum_attempt=2),
-            wait_strategy=wait_strategies.Fixed(wait_time=0),
-            handled_exceptions=self.handled_exceptions)
+        retry_policy = RetryPolicyBuilder() \
+            .with_stop_strategy(stop_strategies.AfterAttempt(maximum_attempt=2)) \
+            .with_wait_strategy(wait_strategies.Fixed(wait_time=0)) \
+            .continue_on_exception(MockException) \
+            .build()
         retry_policy.execute(callable=self.service.call_and_raise)
         assert_greater(self.service.call_count, 1)